* src/graph/graph.cc (GraphInterface::LabelParallelEdges): use long instead of size_t for parallel edges property.
* src/graph/graph_properties.cc: call size_t by "size_t" not "long", since those are not the same type. This removes a bug where it is impossible to create/edit properties with type long.

* src/graph/graph.cc (GraphInterface::LabelParallelEdges): use vector_property_map instead of HashedDescriptorMap.


git-svn-id: https://svn.forked.de/graph-tool/trunk@120 d4600afd-f417-0410-95de-beed9576f240

+++ src/graph/graph_properties.cc
@@ -37,10 +37,10 @@ using namespace graph_tool;
 namespace graph_tool
 {
 // global property types
-const char* type_names[] = {"boolean", "int", "long", "long", "float", "double", "string"};
+const char* type_names[] = {"boolean", "int", "long", "size_t", "float", "double", "string"};
 
 // scalar types
-const char* scalar_names[] = {"boolean", "int", "long", "long", "float", "double"};
+const char* scalar_names[] = {"boolean", "int", "long", "size_t", "float", "double"};
 }
 
 //==============================================================================
